counts collections + resource backup

+++ registry-ws/src/main/java/org/gbif/registry/persistence/mapper/CollectionMapper.java
@@ -26,10 +26,10 @@ public interface CollectionMapper extends TaggableMapper, IdentifiableMapper {
 
   List<Collection> search(@Nullable @Param("query") String query, @Nullable @Param("page") Pageable page);
 
-  int count();
+  long count();
 
-  int countWithFilter(@Nullable @Param("query") String query);
+  long countWithFilter(@Nullable @Param("query") String query);
 
-  int countByInstitution(@Param("institutionKey") UUID institutionKey);
+  long countByInstitution(@Param("institutionKey") UUID institutionKey);
 
 }
